+++ jdto/src/main/java/org/jdto/impl/BeanClassUtils.java
@@ -24,9 +24,10 @@ import org.slf4j.LoggerFactory;
 
 /**
  * DO NOT USE THIS CLASS AS IS NOT PART OF JDTO PUBLIC API <br />
- * 
- * Utility class for creating instances of classes by reflection. The methods
- * on this class handle reflective creation exceptions and log them properly.
+ *
+ * Utility class for creating instances of classes by reflection. The methods on
+ * this class handle reflective creation exceptions and log them properly.
+ *
  * @author Juan Alberto Lopez Cavallotti
  */
 public class BeanClassUtils {
@@ -34,9 +35,10 @@ public class BeanClassUtils {
     private static final Logger logger = LoggerFactory.getLogger(BeanClassUtils.class);
 
     /**
-     * Create a new instance of a class or log the exception if the class is
-     * not instanceable. This method will rethrow any exception as an unchecked
+     * Create a new instance of a class or log the exception if the class is not
+     * instanceable. This method will rethrow any exception as an unchecked
      * {@link RuntimeException}.
+     *
      * @param <T>
      * @param cls
      * @return a new instance of the type given in the argument.
@@ -54,15 +56,33 @@ public class BeanClassUtils {
      * Create a new instance of a class using a specific constructor or log the
      * exception if the class is not instanceable. This method will rethrow any
      * exception as an unchecked {@link RuntimeException}.
+     *
      * @param <T>
      * @param cls
      * @param constructor
      * @param argValues
-     * @return a new instance of the type given on the argument by using the specified constructor.
+     * @return a new instance of the type given on the argument by using the
+     * specified constructor.
      */
-    public static <T> T createInstance(Class<T> cls, Constructor constructor, ArrayList argValues) {
+    public static <T> T createInstance(Class<T> cls, Constructor constructor, List argValues) {
+        Object[] args = argValues.toArray();
+        return createInstance(cls, constructor, args);
+    }
+
+    /**
+     * Create a new instance of a class using a specific constructor or log the
+     * exception if the class is not instanceable. This method will rethrow any
+     * exception as an unchecked {@link RuntimeException}.
+     *
+     * @param <T> the type of the instance to create
+     * @param cls the class which represents the type of the instance to create.
+     * @param constructor the constructor to use.
+     * @param args an array of the constructor arguments.
+     * @return a new instance of the type given on the argument by using the
+     * specified constructor.
+     */
+    public static <T> T createInstance(Class<T> cls, Constructor constructor, Object... args) {
         try {
-            Object[] args = argValues.toArray();
             return (T) constructor.newInstance(args);
         } catch (Throwable t) {
             logger.error("Could not create bean instance of class" + cls.toString(), t);
@@ -71,18 +91,18 @@ public class BeanClassUtils {
     }
 
     /**
-     * Create a new collection of the given type. <br />
-     * If the given type cannot be instantiated then the following applies: <br />
-     * If the given collection is a List, return a new ArrayList. <br />
-     * If the given collection is a Set, return a new HashSet. <br />
-     * If none of the previous, then return a new array list. <br />
-     * 
+     * Create a new collection of the given type. <br /> If the given type
+     * cannot be instantiated then the following applies: <br /> If the given
+     * collection is a List, return a new ArrayList. <br /> If the given
+     * collection is a Set, return a new HashSet. <br /> If none of the
+     * previous, then return a new array list. <br />
+     *
      * @param <T> The type of the collection to be returned.
      * @param collectionClass the type of the collection to be instantiated.
      * @return a new instance of the given collection.
      */
     public static <T extends Collection> T createCollectionInstance(Class<?> collectionClass) {
-        
+
         T ret = null;
         try {
             ret = (T) collectionClass.newInstance();
@@ -101,6 +121,7 @@ public class BeanClassUtils {
 
     /**
      * Check if the class has a default constructor.
+     *
      * @param cls
      * @return true if the type has default constructor, false if not.
      */
@@ -120,6 +141,7 @@ public class BeanClassUtils {
 
     /**
      * Try to find a class out of a string or return null;
+     *
      * @param type
      * @return the Class object represented on the given parameter.
      */
@@ -138,11 +160,13 @@ public class BeanClassUtils {
     }
 
     /**
-     * Find a constructor in the class for the argument types. This method converts
-     * any checked exception in unchecked exception.
+     * Find a constructor in the class for the argument types. This method
+     * converts any checked exception in unchecked exception.
+     *
      * @param cls the class whos constructor will be found.
      * @param types the types of the parameters of the constructor.
-     * @return the constructor of the given class which has the given parameter types.
+     * @return the constructor of the given class which has the given parameter
+     * types.
      */
     public static Constructor safeGetConstructor(Class cls, Class[] types) {
         try {
@@ -152,21 +176,23 @@ public class BeanClassUtils {
             throw new RuntimeException("Error while trying to find constructor for class " + cls.getCanonicalName(), ex);
         }
     }
-    
+
     /**
      * Invoke a static method on a class and return the result without throwing
      * exceptions. Just Silently return null if something bad happens.
+     *
      * @param cls the class in which to invoke the static method.
      * @param methodName the name of the method to be invoked.
      * @param args the arguments that will be sent to the method.
-     * @return the result of the invocaiton (which may be null) or null if something goes wrong.
+     * @return the result of the invocaiton (which may be null) or null if
+     * something goes wrong.
      * @since 1.2
      */
     public static Object invokeStaticMethod(Class cls, String methodName, Object[] args) {
         try {
             return MethodUtils.invokeStaticMethod(cls, methodName, args);
         } catch (Exception ex) {
-            logger.error("Got exception while trying to invoke static method: "+methodName + " on  class: "+cls.getCanonicalName(), ex);
+            logger.error("Got exception while trying to invoke static method: " + methodName + " on  class: " + cls.getCanonicalName(), ex);
             return null;
         }
     }